refactor: 테스트를 패키지 구조로 분리
