+++ src/test/java/janggi/domain/piece/ChariotTest.java
@@ -1,11 +1,9 @@
-package janggi.domain;
+package janggi.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import janggi.domain.Position;
 import janggi.domain.board.Board;
-import janggi.domain.piece.Chariot;
-import janggi.domain.piece.Piece;
-import janggi.domain.piece.Soldier;
 import janggi.domain.team.TeamType;
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
