+++ src/test/java/janggi/domain/piece/ElephantTest.java
@@ -1,11 +1,9 @@
-package janggi.domain;
+package janggi.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import janggi.domain.Position;
 import janggi.domain.board.Board;
-import janggi.domain.piece.Horse;
-import janggi.domain.piece.Piece;
-import janggi.domain.piece.Soldier;
 import janggi.domain.team.TeamType;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -15,13 +13,13 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-public class HorseTest {
+public class ElephantTest {
 
     @Nested
     @DisplayName("이동 가능한 위치 계산 테스트")
     class CalculateMovablePositions {
 
-        static Piece horse;
+        static Piece elephant;
         static Piece enemy1;
         static Piece enemy2;
         static Piece enemy3;
@@ -34,7 +32,7 @@ public class HorseTest {
 
         @BeforeEach
         void setUp() {
-            horse = new Horse(TeamType.RED);
+            elephant = new Elephant(TeamType.RED);
             enemy1 = new Soldier(TeamType.BLUE);
             enemy3 = new Soldier(TeamType.BLUE);
             enemy4 = new Soldier(TeamType.BLUE);
@@ -47,48 +45,49 @@ public class HorseTest {
         }
 
         @Test
-        @DisplayName("마는 기물을 뛰어넘을 수 없다.")
+        @DisplayName("상은 기물을 뛰어넘을 수 없다.")
         void test1() {
-            positionPieceMap.put(Position.valueOf(6, 4), horse);
-            positionPieceMap.put(Position.valueOf(6, 3), ally1);
+            positionPieceMap.put(Position.valueOf(6, 4), elephant);
+            positionPieceMap.put(Position.valueOf(4, 1), ally1);
             positionPieceMap.put(Position.valueOf(5, 4), enemy1);
             positionPieceMap.put(Position.valueOf(7, 4), ally2);
-            positionPieceMap.put(Position.valueOf(6, 5), ally3);
+            positionPieceMap.put(Position.valueOf(5, 6), ally3);
+            positionPieceMap.put(Position.valueOf(8, 1), ally4);
+            positionPieceMap.put(Position.valueOf(7, 6), enemy2);
             List<Position> expected = List.of();
 
             Board board = new Board(positionPieceMap);
-            List<Position> actual = horse.calculateMovablePositions(Position.valueOf(6, 4), board);
+            List<Position> actual = elephant.calculateMovablePositions(Position.valueOf(6, 4), board);
 
             assertThat(actual).hasSameElementsAs(expected);
         }
 
         @Test
-        @DisplayName("마는 직선 한 칸, 대각선 한 칸을 가서 기물을 잡을 수 있다.")
+        @DisplayName("상은 직선 한 칸, 대각선 두 칸을 가서 기물을 잡을 수 있다.")
         void test2() {
-            positionPieceMap.put(Position.valueOf(6, 4), horse);
-            positionPieceMap.put(Position.valueOf(5, 6), enemy1);
-            positionPieceMap.put(Position.valueOf(7, 2), enemy2);
-            List<Position> expected = List.of(Position.valueOf(4, 3), Position.valueOf(4, 5),
-                    Position.valueOf(5, 6), Position.valueOf(7, 6),
-                    Position.valueOf(8, 3), Position.valueOf(8, 5), Position.valueOf(7, 2),
-                    Position.valueOf(5, 2));
+            positionPieceMap.put(Position.valueOf(6, 4), elephant);
+            positionPieceMap.put(Position.valueOf(3, 2), enemy1);
+            positionPieceMap.put(Position.valueOf(4, 7), enemy2);
+            List<Position> expected = List.of(Position.valueOf(3, 2), Position.valueOf(3, 6), Position.valueOf(4, 1),
+                    Position.valueOf(4, 7), Position.valueOf(8, 1), Position.valueOf(8, 7), Position.valueOf(9, 2),
+                    Position.valueOf(9, 6));
 
             Board board = new Board(positionPieceMap);
-            List<Position> actual = horse.calculateMovablePositions(Position.valueOf(6, 4), board);
+            List<Position> actual = elephant.calculateMovablePositions(Position.valueOf(6, 4), board);
 
             assertThat(actual).hasSameElementsAs(expected);
         }
 
         @Test
-        @DisplayName("마는 장기판 밖으로 이동할 수 없다.")
+        @DisplayName("상은 장기판 밖으로 이동할 수 없다.")
         void test3() {
-            positionPieceMap.put(Position.valueOf(1, 1), horse);
+            positionPieceMap.put(Position.valueOf(1, 1), elephant);
             positionPieceMap.put(Position.valueOf(1, 2), ally1);
             positionPieceMap.put(Position.valueOf(2, 1), ally3);
             List<Position> expected = List.of();
 
             Board board = new Board(positionPieceMap);
-            List<Position> actual = horse.calculateMovablePositions(Position.valueOf(1, 1), board);
+            List<Position> actual = elephant.calculateMovablePositions(Position.valueOf(1, 1), board);
 
             assertThat(actual).hasSameElementsAs(expected);
         }
